+++ kentikapi/internal/api_endpoints/types.go
@@ -1,0 +1,5 @@
+package api_endpoints
+
+import "github.com/kentik/community_sdk_golang/kentikapi/models"
+
+type ResourceID = models.ID
