+++ kentikapi/internal/validation/validate.go
@@ -1,4 +1,4 @@
-package utils
+package validation
 
 import (
 	"fmt"
